moving convert date to hours
